+++ CHANGELOG.md
@@ -12,10 +12,12 @@ For a full diff see [`675601b...master`][675601b...master].
 
 * Added `Holder` ([#7]), by [@localheinz]
 * Added `Year` ([#10]), by [@localheinz]
+* Added `Range` ([#12]), by [@localheinz]
 
 [675601b...master]: https://github.com/ergebnis/license/compare/675601b...master
 
 [#7]: https://github.com/ergebnis/license/pull/7
 [#10]: https://github.com/ergebnis/license/pull/10
+[#12]: https://github.com/ergebnis/license/pull/12
 
 [@localheinz]: https://github.com/localheinz